Suppress warnings
While I think warnings should be kept for the script itself, we really don’t want warnings in the dependencies since we cannot do anything about it.

Sadly cannot do one or the other.

When / if we make ourselves depend on libSwiftPM then we can choose which warnings to show and have much more configuration here.