Add support for local dependencies
If the dependency is not scp or a github dependency with
the `@` symbol prefix, we check if the string path is valid in the
user's file system using the Path dependency. This way, we still
allow github dependencies without the `@` prefix.

If determined to be a local dependency, a new `DependencyName`
enum case holds the dependency as a Path, which facilitates
providing the string for `Package.swift` and when decoding.
Moreover, this enum case is used to provide the correct `packageline`.

The related test makes use of `Path.home` to have a directory
that is valid for testing wherever the test is run.

Also, updates README.