try to fix Swift the 4.2 build
CryptoSwift only builds for swift 4.2 using a much older version (tag 0.15.0, branch swift42).

There doesn’t seem to be a good way to use specific tags in a dependency specification for specific swift toolchains.

I found two approaches documented in https://github.com/apple/swift-package-manager/blob/master/Documentation/Usage.md#handling-version-specific-logic

1 - make the project you depend on have tags that specify the swift version (e.g., “0.15.0@swift-4.2”).

2 - clone your Package file and make one specific to the older versions (“Package@swift-4.2.swift”, “Package@swift-5.0.swift”).

It would be more useful if there was a way to specify this on the package dependency specification line directly, but there isn’t that I see.

I’ve requested CryptoSwift adopt the tag convention in (1) above, but in the mean time this sets it to 0.15.0 for all swift versions.

Unfortunately, going back to CryptoSwift 0.15.0 means we lose it’s fix for the too generic “Tests” test target name.

To fix that this commit updates to version 1.2.0 of the Version package which contains a similar fix.