fixes #132 - spaces in module paths
This works, and all tests pass, but I’m not that happy with it.  Seems like there’s a smarter regular expression change that can reduce the extra trimming of spaces off the end of things which I had to add.

Also the encoding dance for URLComponents isn’t as concise as it could, perhaps, be?

Works though! :)